requirements.txt
allure-pytest==2.14.0
allure-python-commons==2.14.0
annotated-types==0.7.0
appnope==0.1.4
asgiref==3.7.2
asttokens==2.4.1
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.6.2
blis==1.0.1
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
coverage==7.6.12
cycler==0.12.1
cymem==2.0.8
DAWG-Python==0.7.2
debugpy==1.8.7
decorator==5.1.1
Django==4.2.2
django-bootstrap5==23.4
django-debug-toolbar==4.3.0
executing==2.1.0
Flask==2.3.3
fonttools==4.53.0
gunicorn==22.0.0
h11==0.14.0
idna==3.6
iniconfig==2.1.0
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markovify==0.9.4
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
murmurhash==1.0.10
nest-asyncio==1.6.0
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymorphy3==2.0.2
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.3
pytz==2024.2
pyzmq==26.2.0
requests==2.31.0
rich==13.9.2
ru_core_news_md @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_md-3.8.0/ru_core_news_md-3.8.0-py3-none-any.whl#sha256=43f456d40e4b70726874e01ce01fce32f50089ba98a2660292e9a6c63a06ebe9
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.29.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
text-unidecode==1.3
thinc==8.3.2
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==2.3.7
wordcloud==1.9.4
wrapt==1.16.0
wsproto==1.2.0
